--- a/core/queryresultio/binary/src/main/java/org/eclipse/rdf4j/query/resultio/binary/BinaryQueryResultWriter.java
+++ b/core/queryresultio/binary/src/main/java/org/eclipse/rdf4j/query/resultio/binary/BinaryQueryResultWriter.java
@@ -88,6 +88,7 @@ public class BinaryQueryResultWriter extends AbstractQueryResultWriter implement
 	 *--------------*/
 
 	public BinaryQueryResultWriter(OutputStream out) {
+		super(out);
 		this.out = new DataOutputStream(out);
 	}