--- a/core/rio/n3/src/main/java/org/eclipse/rdf4j/rio/n3/N3Writer.java
+++ b/core/rio/n3/src/main/java/org/eclipse/rdf4j/rio/n3/N3Writer.java
@@ -50,6 +50,7 @@ public class N3Writer extends AbstractRDFWriter implements RDFWriter {
 	 * @param baseIRI used to relativize IRIs to relative IRIs.
 	 */
 	public N3Writer(OutputStream out, ParsedIRI baseIRI) {
+		super(out);
 		ttlWriter = new TurtleWriter(out, baseIRI);
 	}