--- a/tools/console/src/main/java/org/eclipse/rdf4j/console/util/ConsoleRDFWriter.java
+++ b/tools/console/src/main/java/org/eclipse/rdf4j/console/util/ConsoleRDFWriter.java
@@ -12,10 +12,8 @@ import java.util.Map;
 
 import org.eclipse.rdf4j.console.ConsoleIO;
 import org.eclipse.rdf4j.console.Util;
-
 import org.eclipse.rdf4j.model.Statement;
 import org.eclipse.rdf4j.query.QueryResultHandlerException;
-
 import org.eclipse.rdf4j.rio.RDFFormat;
 import org.eclipse.rdf4j.rio.RDFHandlerException;
 import org.eclipse.rdf4j.rio.helpers.AbstractRDFWriter;