GitHub issue resolved: #400 <!-- add a Github issue number here, e.g #123. This line 
                              automatically closes the issue when the PR is merged --> 

Briefly describe the changes proposed in this PR:

* Allow ProtocolSession to bypass handler and stream response data directly to Outputstream if applicable
* added getOutputStream method to make wrapped outputstream accessible to protocol session.