@jeenbroekstra 2020-03-16
I've ran some ad-hoc local performance tests, manually, and am seeing significant performance improvements when this bypass is used for queries with large result sets. I tested with a construct query that returned ~100,000 triples, and  full evaluation / result handling time went down from ~3000ms to ~2000ms. On a smaller result (~4000 triples), handling time went down from ~600ms to ~200ms.