@hmottestad 2020-03-16
LGTM. Just to let you know, this has now become a part of java 9: https://dzone.com/articles/transferring-inputstream-to-outputstream-in-jdk-9